% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2xlsx.R
\name{pdf2xlsx}
\alias{pdf2xlsx}
\title{Convert pdf (data table) to xlsx (excel)}
\usage{
pdf2xlsx(input_pdf)
}
\arguments{
\item{input_pdf}{Character, a name of pdf file}
}
\value{
xlsx file contains all allocated projects
}
\description{
Getting all allocated projects in a given province
}
\examples{
input_pdf <- system.file("extdata","0150000s.pdf",package = "pdf2xlsx")
browseURL(input_pdf)
pdf2xlsx(input_pdf)

}
